banned users are automatically logged out when trying to access their profile page or a challenge page
