+++ src/components/utils/BannedUserPage.js
@@ -1,4 +1,8 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import {
+  fetchFindIfUserBannedByUsername
+} from '../../features/user/userSlice';
 
 // IMAGES
 import Derp from '../../img/Derp.png'
@@ -8,6 +12,21 @@ import Derp from '../../img/Derp.png'
 // ----------------------------------------------------------------------------------
 
 const BannedUserPage = () => {
+  // State
+  const dispatch = useDispatch();
+
+  // Finds out if the logged in user is banned and forces them to sign out immediately upon navigating to their profile page
+  useEffect(() => {
+    if (localStorage.getItem('username')) {
+      dispatch(fetchFindIfUserBannedByUsername(localStorage.getItem('username')))
+        .then(res => {
+          if (res.payload === true) {
+            localStorage.clear()
+          }
+        })
+    }
+  }, [])
+
   return (
     <div className='text-center text-white'>
       <img
